Use sysfs to locate devices nodes
Changes made during the v4l2_subdev API review renamed the /dev/subdev*
device nodes to /dev/v4l-subdev*. Instead of hardcoding the names in the
media-ctl application, use sysfs to retrieve the kernel device name and
assume the device node name is identical.

This isn't a perfect solution, we should really use libudev, but it will
do for now.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>